refactor: 폴더 이동 및 구조 정리
